Mechanics of submitting to InfluxDB in place
This is just from the Roz end of things... because I get lots of:
     Fatal Connection Error!
     Is InfluxDB running?
     INFLUXDB ERROR. Check above for more details :(

That's okay, because the InfluxDB credentials and tablename in the .conf
file are just dummy placeholders at the moment.

I will probably need to do something about array() entries in the dict to be
committed to the InfluxDB, but I'll worry about that once we have a functional
trial database to play with.

	modified:   roz/database_manager.py
	modified:   roz/process_calibrations.py