Compute a useable "flatness" statistic
Spent the day working toward a "flatness" statistic for the calibration frames.
Ended up with two statistics, one for the planar tilt of the frame, and one for
the quadric curvature of the frame.

Flatness statistics are based on how fast the fit surface changes with respect
to the standard deviation in the image.  Values close to zero are very flat;
values around 1 indicate a change on the order of the stddev across the minumum
dimension of the frame; values >> 1 indicate unflat frames.

	modified:   roz/database_manager.py
	modified:   roz/process_calibrations.py
	modified:   roz/utils.py