Set up the HTML table to be uploaded to Confluence
Create the HTML table that will be uploaded to Confluence, containing the
ongoing analysis information of LMI flat field frames.

There's a lot of CSS tricks and BeautifulSoup tags in here, but it seems to
function as intended.  I expect the code added here will be modified and
simplified at some point in the future.

	modified:   roz/confluence_updater.py