Un-hard-code more stuff
Moved the section headers for the HTML table into the XML VOTable file.  In this
way, when/if filters are added to the LMI complement, we need to simply adjust
the XML VOTable file (lmi_filter_table.xml), and the python code will happily
create the new HTML table as intended.

	modified:   lmi_filter_table.xml
	modified:   roz/confluence_updater.py