Minor substantive changes
After spending WAAAAAAAY to much time trying to figure out how to get Confluence
to only keep the last N versions of the table (and failing), just going to keep
everything for now.

The issue seems to be on Confluence's end... digging into the API, it is
properly sending the delete commands, but the internals of Confluence on the
server seem to (unsurprisingly) mangle things up.

Oh well.  Onward!

	modified:   roz/analyze_alert.py
	modified:   roz/confluence_updater.py
	modified:   roz/main_driver.py
	modified:   roz/process_calibrations.py