Change from keyring to ligmos for containing Confluence login info
Use the ligmos infrastructure for holding the Confluence login information,
and Confluence space/page for the LMI Filter Information table.

This commit has the effect of removing the need for keyring, as well as
un-hard-coding the Confluence space/page information from the python source.

The TEMPLATE file for confluence.conf is included here, with the actual
configuration file containing the login information formerly stored in
keyring.

	new file:   config/confluence.conf-TEMPLATE
	modified:   roz/confluence_updater.py