Confluence HTML table has dynamically updated values!!!
The HTML table now has dynamically changing values for the latest flats.

The dynamic portion is saved as a FITS bintable in data/ and updated with each
run, provided the flats in the current run are newer than what was already in
the table.

Lots of ancillary changes to make this happen.

Still to do: make a PNG of the 'latest' flat, and upload it somehow/somewhere.

	modified:   .gitignore
	modified:   data/lmi_filter_table.xml
	modified:   roz/confluence_updater.py
	modified:   roz/database_manager.py
	modified:   roz/gather_frames.py
	modified:   roz/main_driver.py
	modified:   roz/process_calibrations.py
	modified:   roz/utils.py