Create PNG images for newly processed 'Latest' Flats
This commit adds the creation of the flat-field PNG images, the uploading of
same to Confluence, and editing the HTML LMI Filter Information table to
include a hyperlink to the new PNG image.

TODO:  Maybe figure out if a previous PNG image(s) is attached to the Confluence
page and delete it so as to not foul up the page with hundreds (thousands?) of
attachements over time.

	modified:   .gitignore
	modified:   roz/confluence_updater.py
	modified:   roz/graphics_maker.py
	modified:   roz/utils.py
	new file:   thumbnails/README