All LMI filter tables respond to changes in lmi_filter_table.[ecsv,xml]
If a filter is added to the LMI complement at some point (or one removed), only
the lmi_filter_table.[ecsv,xml] file need be changed (along with the section
heading placements in lmi_table_sechead.ecsv or the second table in
lmi_filter_table.xml).  The dynamically saved table portion and the published
HTML table will accurately reflect the current state of this master list.

Had to do some dancing around with the AstroPy tables to make this happen, but
it seems to work for both adding and removing filters from the table.

	modified:   .gitignore
	modified:   roz/confluence_updater.py
	modified:   roz/utils.py