Add `dumbwaiter` to carry data from wherever to a processing location
The major addition in this commit is the `dumbwaiter` for transferring the
desired files for processing from the (presumed remote) directory to a local
processing directory.  This will enable using Roz with NAS-mounted directories,
as it will be in production.

Relatedly, this commit exploits a feature of the LIGMOS library where multiple
configuration tables may be contained within a single file.  Hence, the removal
of confluence.conf-TEMPLATE and dbconfig.conf-TEMPLATE and the addition of
roz.conf-TEMPLATE.

Lots of conformal changes to make these major things happen.

	deleted:    config/confluence.conf-TEMPLATE
	deleted:    config/dbconfig.conf-TEMPLATE
	new file:   config/roz.conf-TEMPLATE
	modified:   roz/confluence_updater.py
	modified:   roz/database_manager.py
	modified:   roz/gather_frames.py
	modified:   roz/main_driver.py
	modified:   roz/send_alerts.py
	modified:   roz/utils.py