Changed `dumbwaiter` function into `class Dumbwaiter`; make tarball
Converted the `dumbwaiter` function into a class.  The purpose of this change
was to hold onto the various pieces of information more easily and to call
class methods with ease.

Added the first part of the cold_storage() method: creating a compressed tarball
of the collected calibration frames.  The next piece of cold_storage() will be
to upload that tarball to the appropriate server and directory.

	modified:   roz/gather_frames.py
	modified:   roz/main_driver.py
	modified:   roz/send_alerts.py
	modified:   roz/utils.py