Bug fixin' while running on data from `dctobs-nas` (over the network)
Various places where the path was a crucial missing piece -- Roz development
up to this point was done with the code running right out of a data directory.
This test is pulling from remote directories one at a time.

Various small bug fixes to address various code crashes.

Biggest change: PNG images are created and uploaded for any binning scheme, but
the countrate / exptime columns of the LMI Filter Information table are only
updated for 2x2 binning.  The rationale for this is that the flat LOOKS the same
regardless of the binning, and since the PNG does not reference counts or
binning, it should be equally useful.

	modified:   .gitignore
	modified:   data/lmi_filter_table.ecsv
	modified:   roz/confluence_updater.py
	modified:   roz/database_manager.py
	modified:   roz/gather_frames.py
	modified:   roz/graphics_maker.py
	modified:   roz/main_driver.py
	modified:   roz/process_calibrations.py
	modified:   roz/send_alerts.py
	modified:   roz/utils.py