Further fixes to deal with empty bias list
Exploit features of ccdp.ImageFileCollection() to return an empty IFC in the
event of no bias frames (at all).  This is a further edge case from what was
fixed in 351e917f446d974f7f38392354c80546c43c061c.  In that case, there were
multiple binning schemes in a given night -- but not all of which contained
concomitant biases.  In the present commit, this fixes the case when there are
no bias frames at all in the given directory.

	modified:   roz/gather_frames.py
	modified:   roz/process_calibrations.py