Deal with the situation of multiple amplifiers
Most of the time, a single amplifier is used to read out the CCD.  Sometimes,
however, a user chooses to use multiple amplifiers for performance.  This broke
the `base_metadata_dict()` function, causing crash.

This commit adds a `parse_lois_ampids()` function to deal with this issue,
which retains the previous behavior for the single-amplifier case, but adds
additional parsing for multiple-amplifier cases.

	modified:   roz/alerting.py
	modified:   roz/main_driver.py
	modified:   roz/process_frames.py