Successfully ran the LSQ analysis of basic optics model
On a single night's data, produced a 5-parameter optics model for the LDT ASC:
   a0 = -86.24 ±  0.05 deg
   xc = 684.09 ±  0.22 pix
   yc = 484.48 ±  0.30 pix
   F  =   2.33 ±  0.10
   R  = 799.6  ± 30.1

This is now encoded in the module!

What this does is allows the drawing of things like RA/DEC lines, ALT/AZ lines,
and even ecliptic and galactic midplane lines!  Oh, and whatever stars and/or
planets we deem interesting and/or important.  WaHoo!

	modified:   roz/allsky_toolbox.py